+++ native/c/test/zutils.cpp
@@ -100,6 +100,15 @@ string get_random_string(const int length, const bool allNumbers)
   return ret;
 }
 
+std::string get_basename(const std::string& fspath)
+{
+  if (fspath.empty()) return "";
+
+  auto lastSlash = fspath.find_last_of("/\\");
+
+  return (lastSlash == std::string::npos) ? std::string(fspath) : fspath.substr(lastSlash + 1);
+} 
+
 string get_random_uss(const string base_dir)
 {
   static bool seeded = false;